Add files via upload
This report describes the solution of the VLSI problem using combinatorial approach. We
will discuss four optimization techniques to solve the problem - Contraint Programming (CP),
propositional SATisfiability (SAT), Satisfiability Modulo Theories (SMT), and Mixed-Integer Linear
Programming (MIP).
Some of the common constraints are common for all the four techniques-
• Non overlapping and width and height(all models)
• Symmetry breaking in all except MIP
The CP implementation was majorly handled by Shamik Basu and Pratiksha Pratiksha has a small
role in CP implementation. Whereas the MIP was majorly handled by Pratiksha and Shamik has
a supporting role. The other two techniques SAT and SMT were completed by both of us together.
Some major issues while solving the problem with different techniques were-
CP- In CP, we faced issues in implementing the designing of the non overlapping constraint for
the rotation case as we cannot use the global constraint for non overlapping constraint in rotation.
SAT- Encoding the primitive inequalities for constraining the width and height of overall plate and
designing the three literal constraints for non overlapping constraint were difficult to implement in
SAT model.
SMT- In SMT, almost everything is same as SAT so there were no separate issues while implementing
the model. The only addition here is the boolean value for the rotation part.
MIP- While printing the output files for the rotation part of MIP model, the execution became
very slow from instance 19 so we had to proceed with not printing the files for the rotation part.
We were able to obtain till instance 18 but later we had to stop printing the output files to proceed.
The overall working time taken in completing the project was approximately 45
days.